Implement comprehensive type conversion for Default() method
- Add automatic type conversion for default values based on target type
- Support conversions: int→int64, string→int64, bool→string, etc.
- Remove need for explicit casting in tests and examples
- Add comprehensive conversion matrix with proper error handling
- Include range validation for numeric types
- Add extensive test suite for all conversion scenarios
- Fix original bug where .Int64().Default(5) failed with type mismatch

BREAKING CHANGE: Tests now expect automatic conversions instead of type mismatches

+++ api_test.go
@@ -8,7 +8,7 @@ import (
 func TestAPIChanges(t *testing.T) {
 	// Test that new Get[T] API works
 	cfg := New()
-	cfg.Define("PORT").Int64().Default(int64(8080))
+	cfg.Define("PORT").Int64().Default(8080)
 
 	if err := cfg.Execute([]string{"test"}); err != nil {
 		t.Fatalf("Config processing failed: %v", err)
